#3 feat: 영화, 라이브채널, 배너 모델 구현 및 컬렉션뷰셀 데이터 바인딩 코드 수정

+++ assignment/Presentation/Cells/ChannelCollectionViewCell.swift
@@ -112,7 +112,7 @@ extension ChannelCollectionViewCell {
         rank: Int,
         channel: String,
         program: String,
-        viewership: Double
+        viewership: Float
     ) {
         imageView.image = image
         rankLabel.text = "\(rank)"
